# Dynamic parallel cpu DAG scheduler 🚀