Based on the idea of CUDA streams.