@EricLBuehler 2025-04-26
**Before**:

```
cpu_graph_matmul_128x128
                        time:   [533.90 µs 535.85 µs 537.77 µs]
```

**After**

```
cpu_graph_matmul_128x128
                        time:   [388.75 µs 390.19 µs 391.65 µs]
                        change: [-27.543% -27.084% -26.626%] (p = 0.00 < 0.05)
                        Performance has improved.
```


This is ~30% faster!